Remove abstract class LargePermanentFormatter
