--- src/Permanent/LargePermanentFormatter.php
@@ -1,58 +1,0 @@
-<?php
-
-namespace CultuurNet\CalendarSummaryV3\Permanent;
-
-use CultuurNet\CalendarSummaryV3\Translator;
-use IntlDateFormatter;
-
-abstract class LargePermanentFormatter
-{
-    protected $daysOfWeek = array(
-        'monday',
-        'tuesday',
-        'wednesday',
-        'thursday',
-        'friday',
-        'saturday',
-        'sunday'
-    );
-
-    /**
-     * @var IntlDateFormatter
-     */
-    protected $fmtDays;
-
-    /**
-     * @var IntlDateFormatter
-     */
-    protected $fmtShortDays;
-
-    /**
-     * @var Translator
-     */
-    protected $trans;
-
-    public function __construct(string $langCode)
-    {
-        $this->fmtDays = new IntlDateFormatter(
-            $langCode,
-            IntlDateFormatter::FULL,
-            IntlDateFormatter::FULL,
-            date_default_timezone_get(),
-            IntlDateFormatter::GREGORIAN,
-            'EEEE'
-        );
-
-        $this->fmtShortDays = new IntlDateFormatter(
-            $langCode,
-            IntlDateFormatter::FULL,
-            IntlDateFormatter::FULL,
-            date_default_timezone_get(),
-            IntlDateFormatter::GREGORIAN,
-            'EEE'
-        );
-
-        $this->trans = new Translator();
-        $this->trans->setLanguage($langCode);
-    }
-}
