+++ src/Permanent/LargePermanentHTMLFormatter.php
@@ -2,11 +2,62 @@
 
 namespace CultuurNet\CalendarSummaryV3\Permanent;
 
+use CultuurNet\CalendarSummaryV3\Translator;
 use CultuurNet\SearchV3\ValueObjects\Offer;
 use CultuurNet\SearchV3\ValueObjects\OpeningHours;
+use IntlDateFormatter;
 
-class LargePermanentHTMLFormatter extends LargePermanentFormatter implements PermanentFormatterInterface
+class LargePermanentHTMLFormatter implements PermanentFormatterInterface
 {
+    protected $daysOfWeek = array(
+        'monday',
+        'tuesday',
+        'wednesday',
+        'thursday',
+        'friday',
+        'saturday',
+        'sunday'
+    );
+
+    /**
+     * @var IntlDateFormatter
+     */
+    protected $fmtDays;
+
+    /**
+     * @var IntlDateFormatter
+     */
+    protected $fmtShortDays;
+
+    /**
+     * @var Translator
+     */
+    protected $trans;
+
+    public function __construct(string $langCode)
+    {
+        $this->fmtDays = new IntlDateFormatter(
+            $langCode,
+            IntlDateFormatter::FULL,
+            IntlDateFormatter::FULL,
+            date_default_timezone_get(),
+            IntlDateFormatter::GREGORIAN,
+            'EEEE'
+        );
+
+        $this->fmtShortDays = new IntlDateFormatter(
+            $langCode,
+            IntlDateFormatter::FULL,
+            IntlDateFormatter::FULL,
+            date_default_timezone_get(),
+            IntlDateFormatter::GREGORIAN,
+            'EEE'
+        );
+
+        $this->trans = new Translator();
+        $this->trans->setLanguage($langCode);
+    }
+
     public function format(Offer $offer): string
     {
         $output = '';
